+++ NamonaProject_v3/NamonaProject_v3/Model/ClothesModel.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc.TagHelpers.Cache;
 using Microsoft.EntityFrameworkCore;
+using Namona_v3.DTO;
 using NamonaProject_v3_.DTO;
 using NamonaProject_v3_.Persistance;
 
@@ -16,7 +17,7 @@ namespace NamonaProject_v3_.Model
 
         public IEnumerable<AllClothesDto> GetAllClothes()
         {
-            return _context.clothes.Select(x => new AllClothesDto
+            return _context.clothes.Include(x => x.Category).Include(x => x.Gender).Select(x => new AllClothesDto
             {
                 ClothingId = x.ClothingId,
                 ClothingName = x.ClothingName,
@@ -24,62 +25,104 @@ namespace NamonaProject_v3_.Model
                 Color = x.Color,
                 Price = x.Price,
                 Stock = x.Stock,
-                GenderId = x.GenderId,
+                GenderName = x.Gender.GenderType,
+                CategoryName   = x.Category.CategoryName
+            });
+        }
+        public IEnumerable<GenderDto> GetGenders()
+        {
+            return _context.genders.Select(x => new GenderDto
+            {
+                Genderid = x.GenderId,
+                Gendertype = x.GenderType
+            });
+        }
+
+        public IEnumerable<CategoryDto> GetCategories()
+        {
+            return _context.categories.Select(x => new CategoryDto
+            {
+                Id = x.CategoryId,
+                CategoryName = x.CategoryName
             });
         }
         //[Authorize(Roles = "Admin")]
         public async Task ChangeClothingData(ChangeClothingDataDto dto)
         {
             int Id = _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().ClothingId;
+            if (!_context.categories.Any(x => x.CategoryName == dto.Category))
+            {
+                throw new KeyNotFoundException("nincs ilyen kategória");
+            }
+            if (!_context.genders.Any(x => x.GenderType == dto.GenderType))
+            {
+                throw new KeyNotFoundException("nincs ilyen nem");
+            }
+            int CategId = _context.categories.Where(x => x.CategoryName == dto.Category).First().CategoryId;
+            int Genderid = _context.genders.Where(x => x.GenderType == dto.GenderType).First().GenderId;
             using (var trx = _context.Database.BeginTransaction())
             {
                 _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().ClothingName = dto.ClothingName;
-                _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().Collection = dto.Collection;
-                _context.clothes.Include(x => x.Category).Where(x => x.ClothingId == dto.ClothingId).First().Category.CategoryName = dto.Category;
+                _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().Collection = dto.Collection;     
                 _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().Color = dto.Color;
                 _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().Price = dto.Price;
-                _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().GenderId = _context.genders.Where(x => x.GenderType == dto.GenderType).First().GenderId;
+                _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().GenderId = Genderid;
+                _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().CategoryId = CategId;
                 _context.clothes.Where(x => x.ClothingId == dto.ClothingId).First().Stock = dto.Stock;
 
-                _context.SaveChanges();
-                trx.Commit();
+                await _context.SaveChangesAsync();
+                await trx.CommitAsync();
             }
             await Task.CompletedTask;
         }
         //[Authorize(Roles = "Admin")]
         public async Task DeleteClothes(int id)
         {
+            if(!_context.clothes.Any(x => x.ClothingId == id))
+            {
+                throw new KeyNotFoundException();   
+            }
             using (var trx = _context.Database.BeginTransaction())
             {
                 _context.clothes.Remove(_context.clothes.Where(x => x.ClothingId == id).First());
-                _context.SaveChanges();
-                trx.Commit();
+                await _context.SaveChangesAsync();
+                await trx.CommitAsync();
             }
             await Task.CompletedTask;
         }
 
         public async Task AddClothes(AddClothesDto dto)
         {
-            using(var trx = _context.Database.BeginTransaction())
+            if (!_context.categories.Any(x => x.CategoryName == dto.CategoryName))
+            {
+                throw new KeyNotFoundException("nincs ilyen kategória");
+            }
+            if (!_context.genders.Any(x => x.GenderType == dto.GenderName))
+            {
+                throw new KeyNotFoundException("nincs ilyen nem");
+            }
+            int CategId = _context.categories.Where(x => x.CategoryName == dto.CategoryName).First().CategoryId;
+            int Genderid = _context.genders.Where(x => x.GenderType == dto.GenderName).First().GenderId;
+            using (var trx = _context.Database.BeginTransaction())
             {
                 _context.clothes.Add(new Clothes
                 { 
                     ClothingName = dto.ClothingName,
                     Collection = dto.Collection,
-                    CategoryId = dto.CatgeroryId,
-                    GenderId = dto.GenderId,
+                    CategoryId = CategId,
+                    GenderId = Genderid,
                     Stock = dto.Stock,
                     Color = dto.Color,
                     Price = dto.Price,
                     
                 });
-                _context.SaveChanges();
-                trx.Commit();
-               
+                await _context.SaveChangesAsync();
+                await trx.CommitAsync();
+
             }
             await Task.CompletedTask;
         }
-
+        /*
         public IEnumerable<AllClothesDto> FilterClothes(
             string category,
             string collection,
@@ -125,75 +168,105 @@ namespace NamonaProject_v3_.Model
                 Price = x.Price
             }).ToList();
 
-
-        }
-        public IEnumerable<AllClothesDto> FilterClothes2(string category, string collection, string gender, int minprice = 0, int maxprice = 99999999)
+        
+        }*/
+        public IEnumerable<AllClothesDto> FilterClothes2(FilterClothesDto dto)
         {
             var result = _context.clothes
                 .AsQueryable();
-            if (category != null && collection != null && gender != null)
+            if(dto.Minprice == 0)
+            {
+
+            dto.Minprice = _context.clothes.Min(x => x.Price);
+            }if (dto.Maxprice == 999999)
+            {
+                dto.Maxprice = _context.clothes.Max(x => x.Price);
+            }
+            if (dto.Category != null && dto.Collection != null && dto.Gender != null)
             {
                 result =
                 _context.clothes.Include(x => x.Gender).Include(x => x.Category)
-                .Where(x => x.Category.CategoryName == category.ToLower() &&
-                x.Collection == collection.ToLower() &&
-                x.Gender.GenderType == gender);
+                .Where(x => x.Category.CategoryName == dto.Category.ToLower() &&
+                x.Collection == dto.Collection.ToLower() &&
+                x.Gender.GenderType == dto.Gender);
             }
-            else if (category == null && collection != null && gender != null)
+            else if (dto.Category == null && dto.Collection != null && dto.Gender != null)
             {
 
                 result =
                _context.clothes.Include(x => x.Gender)
-               .Where(x => x.Collection == collection.ToLower() &&
-               x.Gender.GenderType == gender);
+               .Where(x => x.Collection == dto.Collection.ToLower() &&
+               x.Gender.GenderType == dto.Gender);
             }
-            else if (category != null && collection == null && gender != null)
+            else if (dto.Category != null && dto.Collection == null && dto.Gender != null)
             {
-                _context.clothes.Include(x => x.Gender).Include(x => x.Category)
-                .Where(x => x.Category.CategoryName == category.ToLower() &&
-                x.Gender.GenderType == gender);
+                result = _context.clothes.Include(x => x.Gender).Include(x => x.Category)
+                .Where(x => x.Category.CategoryName == dto.Category.ToLower() &&
+                x.Gender.GenderType == dto.Gender);
             }
-            else if (category != null && collection != null && gender == null)
+            else if (dto.Category != null && dto.Collection != null && dto.Gender == null)
             {
-                _context.clothes.Include(x => x.Category)
-                .Where(x => x.Category.CategoryName == category.ToLower() &&
-                x.Collection == collection.ToLower() &&
-                x.Gender.GenderType == gender);
+                result = _context.clothes.Include(x => x.Category)
+                .Where(x => x.Category.CategoryName == dto.Category.ToLower() &&
+                x.Collection == dto.Collection.ToLower() &&
+                x.Gender.GenderType == dto.Gender);
             }
 
-            else if (category == null && collection == null && gender != null)
+            else if (dto.Category == null && dto.Collection == null && dto.Gender != null)
             {
 
-                _context.clothes.Include(x => x.Gender)
+                result = _context.clothes.Include(x => x.Gender)
                    .Where(x =>
-                   x.Gender.GenderType == gender);
+                   x.Gender.GenderType == dto.Gender);
             }
-            else if (category != null && collection == null && gender == null)
+            else if (dto.Category != null && dto.Collection == null && dto.Gender == null)
             {
 
-                _context.clothes.Include(x => x.Category)
-                   .Where(x => x.Category.CategoryName == category.ToLower()
+                result = _context.clothes.Include(x => x.Category)
+                   .Where(x => x.Category.CategoryName == dto.Category.ToLower()
                  );
             }
-            else if (category == null && collection != null && gender == null)
+            else if (dto.Category == null && dto.Collection != null && dto.Gender == null)
             {
 
-                _context.clothes
+                result = _context.clothes
                    .Where(x =>
-                   x.Collection == collection.ToLower());
+                   x.Collection == dto.Collection.ToLower());
             }
 
-            return result.Where(x => x.Price > minprice && x.Price < maxprice).Select(x => new AllClothesDto
+            return result.Include(x=> x.Category).Include(x=> x.Gender)
+                .Where(x => x.Price > dto.Minprice && x.Price < dto.Maxprice)
+                .Select(x => new AllClothesDto
             {
                 ClothingId = x.ClothingId,
                 ClothingName = x.ClothingName,
                 Collection = x.Collection,
-                Category = x.Category.CategoryName,
-                GenderId = x.GenderId,
+                CategoryName = x.Category.CategoryName,
+                GenderName = x.Gender.GenderType,
                 Stock = x.Stock,
                 Color = x.Color,
                 Price = x.Price
             });
+        }
+
+        public IEnumerable<AllClothesDto> SearchBar(string text)
+        {
+            return _context.clothes
+                .Include(x => x.Category).Include(x => x.Gender)
+                .Where(x => x.Category.CategoryName.ToLower().Contains(text.ToLower()) ||
+                x.Gender.GenderType.ToLower().Contains(text.ToLower()) ||
+                x.Collection.ToLower().Contains(text.ToLower()))
+                .Select(x => new AllClothesDto
+                {
+                    ClothingId = x.ClothingId,
+                    ClothingName = x.ClothingName,
+                    Collection = x.Collection,
+                    CategoryName = x.Category.CategoryName,
+                    GenderName = x.Gender.GenderType,
+                    Stock = x.Stock,
+                    Color = x.Color,
+                    Price = x.Price
+                });
         }
     }
 }
